GitOps with Image Change Triggers
This is a proposed solution to set up GitOps with OpenShift Builds and
ImageChange triggers. Projects declare their initial manifest in the
`init` directory, and apply subsequent changes to initial manifests via
JSON patches in the `patch` directory. The provided `sync` script runs
regular reconciliations using the patch files in the `patch` directory.